feat(124201): Unifica estado do filtro na listagem
