+++ src/componentes/escolas/Receitas/FormFiltrosAvancados/index.js
@@ -8,7 +8,7 @@ import { mantemEstadoFiltrosUnidade } from "../../../../services/mantemEstadoFil
 
 export const FormFiltrosAvancados = (props) => {
 
-    const {btnMaisFiltros, onClickBtnMaisFiltros, setLista, setBuscaUtilizandoFiltro, iniciaLista, buscaTotaisReceitas, previousPath} = props;
+    const {btnMaisFiltros, onClickBtnMaisFiltros, setLista, setBuscaUtilizandoFiltro, iniciaLista, buscaTotaisReceitas, previousPath, state, setState, initialState} = props;
     
     const tabelaInicial = {
         tipos_receita: [],
@@ -16,17 +16,7 @@ export const FormFiltrosAvancados = (props) => {
         contas_associacao: []
     };
 
-    const initialState = {
-        filtrar_por_termo: "",
-        tipo_receita: "",
-        acao_associacao: "",
-        conta_associacao: "",
-        data_inicio: "",
-        data_fim: "",
-    };
-
     const [tabelas, setTabelas] = useState(tabelaInicial);
-    const [state, setState] = useState(initialState);
 
     useEffect(() => {
         if (!previousPath) return;
@@ -56,8 +46,8 @@ export const FormFiltrosAvancados = (props) => {
 
     const handleSubmit = async (event) => {
         event.preventDefault();
-        let data_inicio = state.data_inicio ? moment(new Date(state.data_inicio), "YYYY-MM-DD").format("YYYY-MM-DD") : null;
-        let data_fim = state.data_fim ? moment(new Date(state.data_fim), "YYYY-MM-DD").format("YYYY-MM-DD") : null;
+        let data_inicio = state.data_inicio ? moment(state.data_inicio, "YYYY-MM-DD").format("YYYY-MM-DD") : null;
+        let data_fim = state.data_fim ? moment(state.data_fim, "YYYY-MM-DD").format("YYYY-MM-DD") : null;
         buscaTotaisReceitas(state.tipo_receita, state.acao_associacao, state.conta_associacao, data_inicio, data_fim);
         const lista_retorno_api = await filtrosAvancadosReceitas(state.filtrar_por_termo, state.tipo_receita, state.acao_associacao, state.conta_associacao, data_inicio, data_fim);
         setLista(lista_retorno_api);
